@@ -1,50 +1,52 @@
 source 'https://rubygems.org'
 
-gem 'rails', '4.2.11.1'
+gem 'rails', '5.2.4.3'
 gem 'rake', '13.0.1'
 
 group :development do
   gem 'brakeman' # It is recommended to always use the latest version.
   gem 'bundler-audit'
-  gem 'rubocop', '~> 0.80.1', require: false
-  gem 'spring', '~> 1.1.2'
-  gem 'web-console', '~> 2.3.0'
+  gem 'listen'
+  gem 'rubocop', '~> 0.89.1', require: false
+  gem 'spring', '~> 2.1.1'
+  gem 'web-console', '~> 3.7.0'
 end
 
 group :development, :test do
-  gem 'byebug', '~> 3.1.2'
-  gem 'dotenv-rails', '~> 2.5.0'
-  gem 'rspec-rails', '~> 3.6.1'
+  gem 'byebug', '~> 11.1.3'
+  gem 'dotenv-rails', '~> 2.7.6'
+  gem 'rspec-rails', '~> 4.0.1'
 end
 
 group :test do
-  gem 'factory_girl_rails', '~> 4.4.1'
+  gem 'factory_girl_rails', '~> 4.9.0' # TODO: Replace with factory_bot_rails
   gem 'pundit-matchers', '~> 1.6.0'
   gem 'rspec_junit_formatter' # Used for formatting CircleCI output
-  gem 'shoulda-matchers', '~> 2.8.0'
+  gem 'shoulda-matchers', '~> 4.3.0'
 end
 
-gem 'bcrypt', '~> 3.1.7'
-gem 'bootstrap', '~> 4.4.1'
-gem 'cocoon', '~> 1.2.9'
-gem 'devise', '~> 4.7.1'
-gem 'font-awesome-sass', '~> 5.5.0'
-gem 'jquery-rails', '~> 3.1.0'
-gem 'kaminari', '~> 1.1.1'
-gem 'kramdown', '~> 1.3.3'
-gem 'local_time', '~> 0.3.0'
+gem 'bcrypt', '~> 3.1.15'
+gem 'bootsnap', '~> 1.4.8'
+gem 'bootstrap', '~> 4.5.2'
+gem 'cocoon', '~> 1.2.14'
+gem 'devise', '~> 4.7.2'
+gem 'font-awesome-sass', '~> 5.13.0'
+gem 'jquery-rails', '~> 4.4.0'
+gem 'kaminari', '~> 1.2.1'
+gem 'kramdown', '~> 2.3.0'
+gem 'local_time', '~> 2.1.0'
 gem 'mini_magick', '~> 4.10.1'
-gem 'mini_racer', '~> 0.2.9'
-gem 'nokogiri', '1.10.9'
-gem 'onebox', '~> 1.8.69'
+gem 'mini_racer', '~> 0.3.1'
+gem 'nokogiri', '1.10.10'
+gem 'onebox', '~> 2.0.2'
 gem 'paperclip', '~> 6.1.0'
-gem 'paper_trail', '~> 10.3.1'
-gem 'pg', '~> 0.17.1'
-gem 'pundit', '~> 2.0.0'
+gem 'paper_trail', '~> 11.0.0'
+gem 'pg', '~> 1.2.3'
+gem 'pundit', '~> 2.1.0'
 gem 'puma', '~> 4.3.5'
-gem 'rails_autolink', '~> 1.1.5'
-gem 'ransack', '~> 1.8.8'
-gem 'sanitize', '~> 5.1.0'
-gem 'sass-rails', '~> 5.0.7'
-gem 'stringex', '~> 2.5.2'
-gem 'uglifier', '>= 2.5.0'
+gem 'rails_autolink', '~> 1.1.6'
+gem 'ransack', '~> 2.3.2'
+gem 'sanitize', '~> 5.2.1'
+gem 'sass-rails', '~> 5.1.0'
+gem 'stringex', '~> 2.8.5'
+gem 'uglifier', '>= 4.2.0'
