@@ -1,4 +1,4 @@
-class Contribution < ActiveRecord::Base
+class Contribution < ApplicationRecord
   belongs_to :contributor_profile
   belongs_to :contributable, polymorphic: true
 end
