@@ -1,3 +1,3 @@
-class QuizAnswer < ActiveRecord::Base
+class QuizAnswer < ApplicationRecord
   belongs_to :quiz_question
 end
