@@ -1,4 +1,4 @@
-class QuizQuestion < ActiveRecord::Base
+class QuizQuestion < ApplicationRecord
   belongs_to :quiz
   has_many :quiz_answers, dependent: :destroy
   accepts_nested_attributes_for(
