@@ -1,6 +1,6 @@
 //= require jquery
 //= require jquery_ujs
 //= require cocoon
-//= require local_time
+//= require local-time
 //= require_self
 //= require_tree .
